requirements.txt
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
diffusers==0.33.1
transformers==4.51.3
accelerate==1.6.0
safetensors==0.5.3
tensorflow_cpu==2.19.0
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
fastapi==0.115.12
uvicorn==0.34.2
uvloop==0.21.0
python-dotenv==1.1.0
boto3==1.38.10
openai==1.77.0
pillow==11.2.1
sentencepiece==0.2.0
huggingface-hub==0.30.2
peft==0.15.2
scipy==1.15.2
numpy==2.1.3
psutil==7.0.0
tqdm==4.67.1
filelock==3.18.0
requests==2.32.3
jinja2==3.1.6
s3transfer==0.12.0
typing_extensions==4.13.2